Fix wrong rail_gn size in rail connect message
The bug prevented the passive side of LDMS to accept multiple rail
connections from the same source due to wrong `rail_gn` in the connect
message. The size of `rail_gn` in the message was incorrectly being
32-bit, too small to receive the actual 64-bit rail_gn, resulting in
`msg.rail_gn` being 0 in the network-byte order + bit reduction.

+++ ldms/src/core/ldms_rail.h
@@ -85,7 +85,7 @@ struct ldms_rail_conn_msg_s {
 	int n_eps; /* number of endpoints */
 	uint32_t idx; /* endpoint index in the rail */
 	int pid;
-	int rail_gn;
+	uint64_t rail_gn;
 
 };
 #pragma pack(pop)
